CodeXR_project/requirements.txt
streamlit>=1.34.0
pydantic>=2.7.0
requests>=2.31.0
openai>=1.30.0
google-generativeai>=0.7.0